[Dynamic Min Liq Filter] Implement and test liquidity pricer, worker and liquidity caching in tokens handler (#310)
* [Dynamic Min Liq Filter] Implement and test liquidity pricer, worker and liquidity caching in tokens handler

* godoc

* remove config changes

* updates

* Update tokens/usecase/tokens_usecase.go

+++ domain/router.go
@@ -80,7 +80,7 @@ type RouterConfig struct {
 
 	// Minimum liquidity capitalization for a pool to be considered in the router.
 	// The denomination assumed is pricing.default-quote-human-denom.
-	MinPoolLiquidityCap int `mapstructure:"min-pool-liquidity-cap"`
+	MinPoolLiquidityCap uint64 `mapstructure:"min-pool-liquidity-cap"`
 
 	// Whether to enable route caching
 	RouteCacheEnabled bool `mapstructure:"route-cache-enabled"`
@@ -123,7 +123,7 @@ type RouterOptions struct {
 	MaxRoutes        int
 	MaxSplitRoutes   int
 	// MinPoolLiquidityCap is the minimum liquidity capitalization required for a pool to be considered in the route.
-	MinPoolLiquidityCap int
+	MinPoolLiquidityCap uint64
 	// The number of milliseconds to cache candidate routes for before expiry.
 	CandidateRouteCacheExpirySeconds int
 	RankedRouteCacheExpirySeconds    int
@@ -139,7 +139,7 @@ type RouterOption func(*RouterOptions)
 
 // WithMinPoolLiquidityCap configures the router options with the min pool liquidity
 // capitalization.
-func WithMinPoolLiquidityCap(minPoolLiquidityCap int) RouterOption {
+func WithMinPoolLiquidityCap(minPoolLiquidityCap uint64) RouterOption {
 	return func(o *RouterOptions) {
 		o.MinPoolLiquidityCap = minPoolLiquidityCap
 	}
